Add solution and project files for Visual Studio 2005
These are basic starter files for building in Visual Studio 2005.
The example and linenoise are set up as separate projects, both in
the linenoise-vs2005.sln solution file.  linenoise builds as a
static library (.lib), and the C runtime is statically linked so
no DLL dependencies.  Configurations are set up for Win32 (x86)
and x64.  No need to use these, but they may be convenient for
getting started looking at the code.  VS 2008 and 2010 should be
able to convert the files, they aren't doing anything fancy.