Added permission loading based on username.
If there's no "-p" on the command-line, look for a file
	DIR/permission/USERNAME.txt
and apply that to the permissions.

This lets us have one configuration directory for a number
of users, and then control which user can do what.

By default, users are allowed to run any commands. It doesn't
make much sense to configure many commands, and then to have
a default permissions saying "deny"